+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Renamed `RpcProtocolStream` to `RpcProtocolBlock` to match rename in the SHV
+  standard.
+
 ### Fixed
 - Reduced performance impact of unpacking bytes read due to unnecessary
   abstraction